Address gdkchan's comment

+++ Ryujinx.Audio/Renderer/Dsp/AdpcmHelper.cs
@@ -75,7 +75,7 @@ namespace Ryujinx.Audio.Renderer.Dsp
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static short GetCoefficientAtIndex(ReadOnlySpan<short> coefficients, int index)
         {
-            if ((uint)index > coefficients.Length)
+            if ((uint)index > (uint)coefficients.Length)
             {
                 Logger.Error?.Print(LogClass.AudioRenderer, $"Out of bound read for coefficient at index {index}");
